fix #14: Burn_after_read. We reupload a file already burned after read, the file is unaccessible
Now db informations on a file is refresh at each upload. Even if the file already exist.

This commit also improve the fix https://github.com/guits/pastefile/pull/66
about the cross filesystem move `Error while saving the db: [Errno 18] Invalid cross-device link`

`shutil.move` take the best of this issue `If the destination is on the current filesystem, then os.rename() is used. Otherwise, src is copied (using shutil.copy2()) to dst and then removed.`

The last modification concern flag for `burn_after_read`. The html post value is now a boolean.

That's allow us to put true or false inside. Instead of taking as true if the string is not empty